Add code coverage verification to builds
This commit adds the Jacoco plugin and configures it to enforce a 95% code
coverage. Reports are also always generated with builds.

We also remove the dependency graph generation step as "dependency graph
enablement is blocked by an Amazon enterprise policy".

Details: https://docs.github.com/en/enterprise-cloud@latest/admin/policies/enforcing-policies-for-your-enterprise/enforcing-policies-for-code-security-and-analysis-for-your-enterprise#enforcing-a-policy-to-manage-the-use-of-dependabot-alerts-in-your-enterprise