Eliminate open-range GET requests
This commit adds support for breaking up open-range GET requests and
fetches data chunk-by-chunk. A very naive implementation of a block
manager is added.

Furthermore, we changed the interface of Object Client
to async to allow for prefetching in the future. The Block Manager is
also written with prefetching in mind. For now, we keep a limited number
of IOBlocks in memory (which is of no use today) which should be very
similar to how prefetching might happen in the future (except rather
than keeping bytes around AFTER they've been read we will prefetch them
to their respective IOBlocks BEFORE they've been read).

There are improvements around testability: we abstract away some SDK
constructs like GetObjectRequest, HeadObjectRequest to limit the
parameter space. This might allow in the future to introduce
property-based tests or fuzz testing and also makes it easier to
prototype 'fake clients' or in-memory models easier.

On top of unit tests we add a new reference test on a full sequential
read and compare with our in-memory model. It seems to pass!

Next steps are to eliminate the byte-by-byte reading and refactor
read() calls into the more efficient read(b, off, len) style.