Simplify local development of integrations
While re-thinking the dependency model in the S3FileIO--SeekableStream
integrations we idenfitifed the need to be able to easily consume new LOCAL
SeekableStream changes.

That is, when I introduce a change in SeekableStream, I should be able to
easily consume this when also working on S3FileIO changes and changes
should propagate seemlessly to "another IDE window".

This can be done easily via Maven local repositories and a gradle plugin.

See https://github.com/amazon-contributing/private-iceberg-staging/pull/20
for how this looks like from another Gradle setup.