Update instructions on how to run micro-benchmarks
While investigating an issue I had to run micro-benchmarks and wanted to
run a subset only. Added the command for this.

The microbenchmarks also ran very long with the default number of
repetitions (15) so lowered this (to 3). This should give us a better
development experience when we implement the sequential prefetcher.