[Hotfix] Remove `join()` statements from PhysicalIO
We observed that benchmark runs can hang due to CompletableFuture `join()`s
being waited on indefinitely.  This change replaces all `join()`s with the
`get(timeout, unit)` equivalent.

The timeouts are configurable and are made part of the
PhysicalIOConfiguration.

We also add regression tests on (the lack of) hanging behaviour on both
the GET and the HEAD path.