[Fix] Resolve dependency conflict in uber JAR
We realised that commons-io was causing depency conflicts when the seekable
stream is built into Hadoop.  For now, we shade this dependency, we may want
to remove it later or lift it into the 'common' module.

While I was here, I moved some reference test dependency descriptions into
the common .toml file.