Update Iceberg integration in CI/CD
After recent syncs with Jack, we updated our Iceberg integration code.

This commit updates to that, given that it is best to test it out early.
We also fixed the checkstyle issues, so the -x checkstyle bit can go now.

Tested the integration JAR on local AQBE runs and it seems to improve some
queries that are regressing currently.