Add a JSON `TelemetryFormat` implementation
This commit implements the recently added `TelemetryFormat` interface to
make it possible to emit telemetry logs in JSON format. This is useful for
developers who want to analyse telemetry logs with tools that understand
structured formats. It also unlocks automation in the future.

This commit adds a new configuration option, `telemetry.format`. When unset,
it defaults to the human readable implementation. When set to `"json"`, it
uses the JSON implementation. We throw up when the config is unknown.

The human-readable (default) format relied on overriding `toString()` which
had to change given that we allow multiple implementations for rendering.
Therefore, we had to move some of the human-readable format's `toString`
methods into the `DefaultTelemetryFormat` -- this should have been part of
the earlier refactoring but I did not notice everything. Consequently, some
tests needed moving too.

We also change the string representation of ranges in the `IOPlan` to be
`start-end` rather than `[start-end]`. This will make logs easier to parse
in the future.