Test Len = 0 and Insufficient Buffer behaviours for positioned reads
According to InputStream documentation Len = 0 should return 0

`If len is zero, then no bytes are read and 0 is returned;`
Following the InputStream implementation, this commit is adding argument checks to
Positioned Reads. This validation is implemented in the SeekableInputStream to ensure
all subclasses behave the same way in the unhappy path.