Disable Iceberg unit tests (#233)
## Description of change

- This PR disables Iceberg unit tests as the latest Iceberg code being
used for running benchmarks had 2 unit tests that are failing - leading
to failure in pushing jars to S3.

- So the unit and integration tests will be enabled in a separate PR
later.

#### Relevant issues
NA

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?
No

#### Does this contribution introduce any new public APIs or behaviors?
No

#### How was the contribution tested?
NA

#### Does this contribution need a changelog entry?
No

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).