[New Physical IO] Added memory manager changes to new physical IO (#288)
## Description of change
This PR adopts the memory manager changes to new physicalIO/

#### Relevant issues
PR History:
- https://github.com/awslabs/analytics-accelerator-s3/pull/286
- https://github.com/awslabs/analytics-accelerator-s3/pull/287

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?
No

#### Does this contribution introduce any new public APIs or behaviors?
No

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Co-authored-by: Erdogan Ozkoca <ozkoca@amazon.com>