Replace DAT with AAL (#300)
## Description of change
This PR replaces all remaining references to DAT with AAL, reflecting
the current and accurate name of the project. These legacy references
have been updated for consistency and clarity.

#### Relevant issues
N/A

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?
No

#### Does this contribution introduce any new public APIs or behaviors?
No

#### How was the contribution tested?
No need to test

#### Does this contribution need a changelog entry?
N/A

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).