New Physical IO changes (#316)
Added iostats get request callback to streamReader (#317)

This PR moves IOStat callback method request from Block to StreamReader.
Needs to be done as part of code rebase

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Telemetry support for physical IO (#318)

This PR adds telemetry measures for StreamReader and BlockManager.