New Physical IO changes (#316)
Added iostats get request callback to streamReader (#317)

This PR moves IOStat callback method request from Block to StreamReader.
Needs to be done as part of code rebase

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

Telemetry support for physical IO (#318)

This PR adds telemetry measures for StreamReader and BlockManager.

Introducing retry policy to new PhysicalIO (#320)

This PR adds the capability of retry to the new PhysicalIO

https://github.com/awslabs/analytics-accelerator-s3/pull/286
https://github.com/awslabs/analytics-accelerator-s3/pull/287
https://github.com/awslabs/analytics-accelerator-s3/pull/288
https://github.com/awslabs/analytics-accelerator-s3/pull/289
https://github.com/awslabs/analytics-accelerator-s3/pull/294
https://github.com/awslabs/analytics-accelerator-s3/pull/316

existing APIs or behaviors?
No

No

Unit tests

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).