Changed default read buffer size to 128KB (#324)
## Description of change
This PR changes default read buffer size to 128KB to have a better
performance

#### Relevant issues
https://github.com/awslabs/analytics-accelerator-s3/pull/286
https://github.com/awslabs/analytics-accelerator-s3/pull/287
https://github.com/awslabs/analytics-accelerator-s3/pull/288
https://github.com/awslabs/analytics-accelerator-s3/pull/289
https://github.com/awslabs/analytics-accelerator-s3/pull/294
https://github.com/awslabs/analytics-accelerator-s3/pull/316
https://github.com/awslabs/analytics-accelerator-s3/pull/320
https://github.com/awslabs/analytics-accelerator-s3/pull/323

#### Does this contribution introduce any breaking changes to the
existing APIs or behaviors?
No

#### Does this contribution introduce any new public APIs or behaviors?
No

#### How was the contribution tested?
Unit tests

#### Does this contribution need a changelog entry?
N/A

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).