fix: update to push snapshot to maven local for iceberg build (#350)
We currently don't use the snapshot build when building the iceberg
artifacts this fixes that. Also adding a manual trigger so we can test
downstream changes and trigger this independent of a PR.

#### Does this contribution need a changelog entry?
- [x] I have updated the CHANGELOG or README if appropriate

---

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license and I agree to the terms of
the [Developer Certificate of Origin
(DCO)](https://developercertificate.org/).

---------

Signed-off-by: stubz151 <stubz151@gmail.com>